requirement.txt
django
djangorestframework
django-cors-headers
drf-yasg
celery
mysqlclient
django-environ

Explanation of Dependencies
django: The core Django framework for building the web application.
djangorestframework: For developing RESTful APIs.
django-cors-headers: To handle Cross-Origin Resource Sharing (CORS) for API requests.
drf-yasg: For generating Swagger/OpenAPI documentation for the APIs.
celery: For handling asynchronous task queuing (to be configured in future milestones).
mysqlclient: The MySQL database adapter for Django to connect to the MySQL database.
django-environ: For managing environment variables securely using .env files.


amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
kombu==5.4.2
mysqlclient==2.2.6
packaging==24.2
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
